mazak: move log csv parsing into its own file
